ENH: Fix warnings MSVC

See merge request iMSTK/iMSTK!609

+++ Examples/TaskGraph/Timing/taskGraphTimingExample.cpp
@@ -77,7 +77,7 @@ static std::shared_ptr<PbdObject>
 makePbdString(
     const std::string& name,
     const Vec3d&       pos,
-    const size_t       numVerts,
+    const int          numVerts,
     const double       stringLength,
     const double       bendStiffness,
     const Color&       color)
